Move socket_vsock_proxy to host/commands
The "common" distinction doesn't apply now that these executables are
built exclusively for the host.
