+++ base/cvd/cuttlefish/host/commands/socket_vsock_proxy/socket_vsock_proxy.cpp
@@ -23,11 +23,11 @@
 #include "absl/log/check.h"
 #include "absl/log/log.h"
 
-#include "cuttlefish/common/frontend/socket_vsock_proxy/client.h"
-#include "cuttlefish/common/frontend/socket_vsock_proxy/server.h"
 #include "cuttlefish/common/libs/fs/shared_fd.h"
 #include "cuttlefish/common/libs/utils/socket2socket_proxy.h"
 #include "cuttlefish/host/commands/kernel_log_monitor/utils.h"
+#include "cuttlefish/host/commands/socket_vsock_proxy/client.h"
+#include "cuttlefish/host/commands/socket_vsock_proxy/server.h"
 #include "cuttlefish/host/libs/config/config_instance_derived.h"
 #include "cuttlefish/host/libs/config/cuttlefish_config.h"
 #include "cuttlefish/host/libs/config/logging.h"
